Creating teacher supports component and adding to groups view
