+++ src/components/teacher/teacher-dashboard.sass
@@ -11,30 +11,6 @@ $tab-height: 20px
     bottom: $double-margin
     color: #000
 
-    .dash-header
-      font-family: inherit
-      font-size: inherit
-      color: black
-      font-weight: bold
-
-      > div
-        margin-bottom: $teacher-margin
-
-      .header-contents
-        display: flex
-
-        div
-          text-align: center
-          margin-left: $teacher-margin
-
-        .date
-          margin-left: $date-margin
-          width: $date-width
-        
-        .section
-          width: $section-width
-
-
     .tabs
       display: flex
       flex-direction: row
