+++ src/components/teacher/teacher-dashboard.tsx
@@ -1,7 +1,5 @@
 import { inject, observer } from "mobx-react";
 import * as React from "react";
-import { TabComponent } from "../tab";
-import { TabSetComponent } from "../tab-set";
 import { BaseComponent, IBaseProps } from "../base";
 import { HeaderComponent } from "../header";
 import { TeacherGroupTabComponent } from "./teacher-group-tab";
@@ -10,7 +8,7 @@ import { TeacherStudentTabComponent } from "./teacher-student-tab";
 import "./teacher-dashboard.sass";
 import { BottomNavComponent } from "../navigation/bottom-nav";
 import { RightNavComponent } from "../navigation/right-nav";
-import { TeacherSupport } from "./teacher-support";
+import { TeacherSupports } from "./teacher-supports";
 
 interface IProps extends IBaseProps {}
 interface IState {
@@ -47,35 +45,12 @@ export class TeacherDashboardComponent extends BaseComponent<IProps, IState> {
         <HeaderComponent isGhostUser={true} />
         <div className="tabbed-area">
           <div className="tab-contents" aria-labelledby={this.getTabId(activeTab)}>
-            { this.renderHeader() }
-            <TeacherSupport time={new Date().getTime()}/>
-            {
-              // Reverse the supports so the newest ones are first + displayed at the top
-              supports.teacherSupports.slice()
-                .filter(support => !support.deleted)
-                .reverse()
-                .map((support, i) => {
-                  return <TeacherSupport support={support} time={support.authoredTime} key={support.key}/>;
-                })
-            }
+            <TeacherSupports supports={supports.teacherSupports.filter(support => !support.deleted)}/>
             {this.renderTabContents()}
           </div>
         </div>
         <BottomNavComponent />
         <RightNavComponent isGhostUser={true} />
-      </div>
-    );
-  }
-
-  private renderHeader() {
-    return (
-      <div className="dash-header">
-        <div className="title">Class Supports:</div>
-        <div className="header-contents">
-          <div className="date">Date</div>
-          <div className="section">Section</div>
-          <div className="content">Message</div>
-        </div>
       </div>
     );
   }
