+++ src/components/teacher/teacher-group-tab.sass
@@ -62,5 +62,5 @@
         padding: $half-padding $padding
         border-radius: 0 $border-radius 0 0
 
-    .content
-      padding: $padding
+  .teacher-supports
+    padding: 5px
