+++ src/components/teacher/teacher-supports.sass
@@ -1,0 +1,25 @@
+@import ../vars
+  
+.teacher-supports
+  .dash-header
+    font-family: inherit
+    font-size: inherit
+    color: black
+    font-weight: bold
+
+    > div
+      margin-bottom: $teacher-margin
+
+    .header-contents
+      display: flex
+
+      div
+        text-align: center
+        margin-left: $teacher-margin
+
+      .date
+        margin-left: $date-margin
+        width: $date-width
+      
+      .section
+        width: $section-width
